Escape characters that glibc doesn't like printed
Rendering a line like "% node gophertest.js" causes a fatal crash on
glibc.  The snprintf implementation doesn't like the "%n" that it sees,
so it prints an error directly to stdout and fails.  The solution is to
escape % signs in a way that snprintf approves.

This behavior has not been tested with any other libc implementation.